Don't use MultiIndex for `get_data()` return
to make graphing easier, since the metric label does not need to
be stored in the data frame since it is already known (it is passed
in as a parameter, or the corresponding label for the ID that is
passed in can be obtained from `get_metrics()`). Also change
`tolist()` to `to_list()` in the test code to be consistent (they are
equivalent).